+++ packages/ui/button/src/Button.stories.tsx
@@ -6,8 +6,8 @@ export default {
   component: Button,
   tags: ['autodocs'],
   argTypes: {
-    left: { control: { type: 'text' } },
-    right: { control: { type: 'text' } },
+    prefix: { control: { type: 'text' } },
+    suffix: { control: { type: 'text' } },
     children: { control: { type: 'text' } }
   }
 }
@@ -17,7 +17,7 @@ export const Default = (args: ButtonProps) => (
     {...args}
     classNames={
       args.classNames ?? {
-        root: 'bg-blue-200 p-2 rounded-md border-1 border-blue-300/50 text-blue-700 hover:bg-blue-300 active:scale-97'
+        root: 'bg-gray-300 rounded-md border-1 border-gray-400/50 text-gray-700 font-medium hover:bg-gray-400 active:scale-97'
       }
     }
   >