+++ packages/ui/button/src/Button.tsx
@@ -3,26 +3,36 @@ import { useNSUI } from '@nsui/provider'
 import { useRipple } from '@nsui/ripple'
 import '@nsui/ripple/ripple.css'
 
-export interface ButtonProps extends React.ButtonHTMLAttributes<HTMLButtonElement> {
-  left?: React.ReactNode
-  right?: React.ReactNode
-  classNames?: { root?: string; children?: string; left?: string; right?: string }
+export interface ButtonProps extends Omit<
+  React.ButtonHTMLAttributes<HTMLButtonElement>,
+  'prefix' | 'className' | 'style'
+> {
+  prefix?: React.ReactNode
+  suffix?: React.ReactNode
+  classNames?: { root?: string; children?: string; prefix?: string; suffix?: string }
   styles?: {
     root?: React.CSSProperties
     children?: React.CSSProperties
-    left?: React.CSSProperties
-    right?: React.CSSProperties
+    prefix?: React.CSSProperties
+    suffix?: React.CSSProperties
   }
-  disableRipple?: boolean
+  isRippleDisabled?: boolean
 }
 
 export const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
-  ({ children, left, right, classNames, styles, disableRipple, onClick, ...props }, ref) => {
-    const { prefixCls, transitionDuration } = useNSUI()
+  (
+    { children, prefix, suffix, classNames, styles, disabled, isRippleDisabled, onClick, ...props },
+    ref
+  ) => {
+    const { global, components } = useNSUI()
+
+    const transitionDuration = components?.button?.transitionDuration ?? global.transitionDuration
+    const isRippleDisabledFinal = isRippleDisabled ?? components?.button?.isRippleDisabled
+
     const { createRipple } = useRipple()
 
     const handleClick: React.MouseEventHandler<HTMLButtonElement> = (e) => {
-      if (!disableRipple) createRipple(e)
+      if (!isRippleDisabledFinal && !disabled) createRipple(e)
       onClick?.(e)
     }
 
@@ -30,7 +40,8 @@ export const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
       <button
         {...props}
         ref={ref}
-        className={`${prefixCls}-btn ${classNames?.root ?? ''}`}
+        disabled={disabled}
+        className={`${global.prefixCls}-btn ${classNames?.root ?? ''}`}
         onClick={handleClick}
         style={{
           display: 'flex',
@@ -38,30 +49,36 @@ export const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
           justifyContent: 'center',
           overflow: 'hidden',
           position: 'relative',
+          gap: '0.2rem',
+          padding: '0.5rem',
           transition: 'all ease-in-out',
+          cursor: disabled ? 'not-allowed' : 'pointer',
           transitionDuration: `${transitionDuration}ms`,
           ...styles?.root
         }}
       >
-        {left && (
-          <span className={`${prefixCls}-btn-left ${classNames?.left ?? ''}`} style={styles?.left}>
-            {left}
+        {prefix && (
+          <span
+            className={`${global.prefixCls}-btn-left ${classNames?.prefix ?? ''}`}
+            style={styles?.prefix}
+          >
+            {prefix}
           </span>
         )}
 
         <span
-          className={`${prefixCls}-btn-content ${classNames?.children ?? ''}`}
+          className={`${global.prefixCls}-btn-content ${classNames?.children ?? ''}`}
           style={styles?.children}
         >
           {children}
         </span>
 
-        {right && (
+        {suffix && (
           <span
-            className={`${prefixCls}-btn-right ${classNames?.right ?? ''}`}
-            style={styles?.right}
+            className={`${global.prefixCls}-btn-right ${classNames?.suffix ?? ''}`}
+            style={styles?.suffix}
           >
-            {right}
+            {suffix}
           </span>
         )}
       </button>